requirements-iker.txt
python-dotenv
accelerate
asttokens
bitsandbytes
certifi
charset-normalizer
colorama
comm
debugpy
decorator
executing
filelock
fsspec
geographiclib
geopy
haversine
huggingface-hub
idna
ipykernel
ipython
ipython_pygments_lexers
jedi
Jinja2
joblib
jupyter_client
jupyter_core
MarkupSafe
matplotlib-inline
mpmath
nest-asyncio
networkx
numpy
packaging
pandas
parso
pillow
platformdirs
prompt_toolkit
psutil
pure_eval
Pygments
python-dateutil
pytz
pywin32
PyYAML
pyzmq
regex
requests
safetensors
scikit-learn
scipy
setuptools
six
stack-data
sympy
threadpoolctl
tokenizers
torch
torchaudio
torchvision
tornado
tqdm
traitlets
transformers
typing_extensions
tzdata
urllib3
wcwidth
streamlit
openpyxl